Moved handlers from rest to diagnostics submodule
